Update jedi-ci for ufo data. (#501)
UFO data was on the old AWS-codebuild based continuous integration pipeline that was outdated. This PR updates UFO data.

Note this PR references the branch @feature/oops-fixes which contains the last batch of updates for the CI overhaul. I will update this to develop once JCSDA-internal/jedi-ci#15 is merged but in the meantime the current branch represents the final ready state of the system that will be rolled out over the next few weeks.